draft jit linear solver class
