+++ include/micm/solver/rosenbrock.inl
@@ -1,3 +1,6 @@
+// Copyright (C) 2023 National Center for Atmospheric Research
+// SPDX-License-Identifier: Apache-2.0
+
 namespace micm
 {
   //
