+++ test/unit/solver/CMakeLists.txt
@@ -13,6 +13,7 @@ create_standard_test(NAME rosenbrock SOURCES test_rosenbrock.cpp)
 create_standard_test(NAME state SOURCES test_state.cpp)
 
 if(ENABLE_LLVM)
-  create_standard_test(NAME jit_rosenbrock SOURCES test_jit_rosenbrock.cpp)
+  create_standard_test(NAME jit_linear_solver SOURCES test_jit_linear_solver.cpp)
   create_standard_test(NAME jit_lu_decomposition SOURCES test_jit_lu_decomposition.cpp)
+  create_standard_test(NAME jit_rosenbrock SOURCES test_jit_rosenbrock.cpp)
 endif()
